Very Big Update: Website (see details)
First i Updated the Fonts adding Nabla, Bungee Spice and Foldit. Secondly i Updated the Header and the beautifull Github footer. And Finaly i added a better Placeholder Matching the Theme of the Game even more, and a half oppacity black box with text in it. And Last but not Least: The Download Links